pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==4.2.5",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
]
